refactor: UsersFactory.java 삭제 등 - refactor(Users): UsersFactory 클래스 삭제로 생성자 인수 형을 String으로 변경 및 컬렉션화 - refactor(User.java): 생성자 유효성체크 순서 가장 먼저하도록 변경
