+++ src/main/java/nextstep/ladder/App.java
@@ -1,12 +1,15 @@
-package nextstep.ladder.domain;
+package nextstep.ladder;
 
-import nextstep.ladder.domain.factory.UsersFactory;
+import nextstep.ladder.domain.Ladder;
+import nextstep.ladder.domain.LadderMaker;
+import nextstep.ladder.domain.LineMakerImpl;
+import nextstep.ladder.domain.Users;
 import nextstep.ladder.ui.InputView;
 import nextstep.ladder.ui.ResultView;
 
 public class App {
     public static void main(String[] args) {
-        Users users = UsersFactory.newInstance(InputView.getUserNames());
+        Users users = new Users(InputView.getUserNames());
         Ladder ladder = new LadderMaker(new LineMakerImpl(users.size())).makeLadder(InputView.getLadderHeight());
         ResultView.printResult(users, ladder);
     }
