+++ src/test/java/nextstep/ladder/domain/UsersTest.java
@@ -1,6 +1,5 @@
 package nextstep.ladder.domain;
 
-import nextstep.ladder.domain.factory.UsersFactory;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -8,7 +7,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 public class UsersTest {
     @Test
     void name() {
-        assertThatThrownBy(() -> UsersFactory.newInstance("aa, aa"))
+        assertThatThrownBy(() -> new Users("aa, aa"))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage("이름이 중복되었습니다.");
     }
